+++ src/sit/tu_varna/bg/project/creator/rectangle/RectangleCreator.java
@@ -39,31 +39,31 @@ public class RectangleCreator extends AbstractCreator {
             return null;
         }
         NamedColors stroke = NamedColors.valueOf(strokeColor.toUpperCase(Locale.ROOT));
-        if (!scanner.hasNextDouble()) {
+        if (!scanner.hasNextInt()) {
             System.out.println("Passed value for stroke width is not a real number!");
             return null;
         }
-        double strokeWidth = scanner.nextDouble();
-        if (!scanner.hasNextDouble()) {
+        int strokeWidth = scanner.nextInt();
+        if (!scanner.hasNextInt()) {
             System.out.println("Passed value for X coordinate is not a real number!");
             return null;
         }
-        double initialX = scanner.nextDouble();
-        if (!scanner.hasNextDouble()) {
+        int initialX = scanner.nextInt();
+        if (!scanner.hasNextInt()) {
             System.out.println("Passed value for Y coordinate is not a real number!");
             return null;
         }
-        double initialY = scanner.nextDouble();
-        if (!scanner.hasNextDouble()) {
+        int initialY = scanner.nextInt();
+        if (!scanner.hasNextInt()) {
             System.out.println("Passed value for width is not a real number!");
             return null;
         }
-        double width = scanner.nextDouble();
-        if (!scanner.hasNextDouble()) {
+        int width = scanner.nextInt();
+        if (!scanner.hasNextInt()) {
             System.out.println("Passed value for height is not a real number!");
             return null;
         }
-        double height = scanner.nextDouble();
+        int height = scanner.nextInt();
         int roundCornerX = 0;
         int roundCornerY = 0;
         if (scanner.hasNextInt()) {
